# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/accounts/{accountId}/spl/buy": {
      "post": {
        "description": "Buy token for some amount of SOL",
        "operationId": "AccountsSolanaController_buyToken",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolanaTransactionEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buy SPL tokens",
        "tags": [
          "spl-token"
        ]
      }
    },
    "/accounts/{accountId}/spl/sell": {
      "delete": {
        "description": "Sell token for some amount of SOL",
        "operationId": "AccountsSolanaController_sellToken",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolanaTransactionEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sell SPL tokens",
        "tags": [
          "spl-token"
        ]
      }
    },
    "/accounts/{accountId}/spl/transfer": {
      "put": {
        "description": "Transfer SPL tokens to defined address",
        "operationId": "AccountsSolanaController_transferToken",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferSplTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SPL tokens transfer started!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolanaTransactionEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transfer SPL tokens",
        "tags": [
          "spl-token"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bundler API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SolanaTransactionEntity": {
        "type": "object",
        "properties": {
          "txHash": {
            "type": "string",
            "description": "The transaction hash"
          }
        },
        "required": [
          "txHash"
        ]
      },
...
      "BuyTokenDto": {
        "type": "object",
        "properties": {
          "solAmount": {
            "type": "number",
            "description": "The amount of SOL to spend",
            "example": 0.001
          },
          "slippagePercent": {
            "type": "number",
            "description": "The slippage percentage",
            "example": 1
          },
          "priorityMicroLamptorsFee": {
            "type": "number",
            "description": "The priority lamports fee",
            "example": 0
          }
        },
        "required": [
          "solAmount"
        ]
      },
      "SellTokenDto": {
        "type": "object",
        "properties": {
          "tokenAmount": {
            "type": "number",
            "description": "The amount of token to sell",
            "example": 10000
          },
          "slippagePercent": {
            "type": "number",
            "description": "The slippage percentage",
            "example": 1
          },
          "skipLimit": {
            "type": "boolean",
            "description": "Skip limit order",
            "example": false,
            "default": false
          },
          "priorityMicroLamptorsFee": {
            "type": "number",
            "description": "The priority lamports fee",
            "example": 0
          }
        },
        "required": [
          "tokenAmount"
        ]
      },
      "TransferSplTokenDto": {
        "type": "object",
        "properties": {
          "percent": {
            "type": "number",
            "description": "Percent of current balance to transfer",
            "example": 100
          },
          "recipient": {
            "type": "string",
            "description": "The address to transfer SPL tokens to"
          },
          "priorityMicroLamptorsFee": {
            "type": "number",
            "description": "The priority lamports fee",
            "example": 0
          }
        },
        "required": [
          "recipient"
        ]
      },
...
    }
  }
}